requirements.txt
altair~=5.3.0
dnspython~=2.6.1
matplotlib~=3.9.0
pandas~=2.2.2
plotly~=5.22.0
requests~=2.32.3
streamlit~=1.36.0
whois~=1.20240129.2
networkx~=3.2.1

numpy~=2.0.0
folium~=0.17.0
tldextract~=5.1.2